+++ src/eval.rs
@@ -1,5 +1,4 @@
-use crate::chess::{Color, Role, Square};
-use crate::position::Position;
+use crate::chess::{Color, Position, Role, Square};
 use crate::search;
 
 pub const INFINITY: i16 = 32_001;
