+++ src/lib.rs
@@ -1,23 +1,20 @@
+pub mod bench;
+pub mod bitboard;
+pub mod chess;
+pub mod limits;
+pub mod movegen;
+pub mod search;
+pub mod tt;
+pub mod uci;
+
 mod eval;
 mod movepicker;
 mod util;
 
-pub mod bench;
-pub mod bitboard;
-pub mod chess;
-pub mod fen;
-pub mod limits;
-pub mod movegen;
-pub mod moves;
-pub mod position;
-pub mod san;
-pub mod search;
-pub mod tt;
-pub mod uci;
-pub mod zobrist;
-
 #[cfg(feature = "datagen")]
 pub mod datagen;
+
+use crate::chess::position::zobrist;
 
 pub fn init() {
     movegen::init_tables();
