+++ src/movegen.rs
@@ -8,7 +8,7 @@ pub use types::{
     BishopType, KingType, KnightType, MoveGen, MoveList, Mover, PawnType, QueenType, RookType,
 };
 
-use crate::position::Position;
+use crate::chess::Position;
 
 mod magic;
 mod magic_gen;
@@ -78,9 +78,9 @@ fn masked_perft(pos: &mut Position, depth: u8) -> usize {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::fen::Fen;
+    use crate::chess::Fen;
+    use crate::chess::fen::STARTPOS;
 
-    const STARTPOS: &str = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
     const KIWIPETE_FEN: &str =
         "r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w KQkq - 0 1";
     const POSITTION_3_FEN: &str = "8/2p5/3p4/KP5r/1R3p1k/8/4P1P1/8 w - - 0 1";
