+++ src/movegen/bishop.rs
@@ -1,9 +1,8 @@
 use types::BishopType;
 
 use crate::bitboard::Bitboard;
-use crate::chess::{Color, Role, Square};
+use crate::chess::{Color, Position, Role, Square};
 use crate::movegen::*;
-use crate::position::Position;
 
 impl Mover for BishopType {
     #[inline]
