+++ src/movegen/types.rs
@@ -1,10 +1,8 @@
 use arrayvec::ArrayVec;
 
 use crate::bitboard::Bitboard;
-use crate::chess::{Color, Role, Square};
+use crate::chess::{Color, Move, Position, Role, Square};
 use crate::movegen::*;
-use crate::moves::Move;
-use crate::position::Position;
 
 pub type MoveList = ArrayVec<FromAndMoves, 18>;
