+++ src/movepicker.rs
@@ -1,10 +1,8 @@
 use arrayvec::ArrayVec;
 
 use crate::bitboard::Bitboard;
-use crate::chess::{Color, Square};
+use crate::chess::{Color, Move, Position, Square};
 use crate::movegen::MoveGen;
-use crate::moves::Move;
-use crate::position::Position;
 
 const CAPTURE_SCORE: i16 = 30_000;
 const KILLER_1_SCORE: i16 = 29_001;
@@ -215,7 +213,7 @@ impl MovePicker {
 
 #[cfg(test)]
 mod test {
-    use crate::fen::Fen;
+    use crate::chess::Fen;
     use crate::movegen::init_tables;
     use crate::zobrist::init_zobrist;
