+++ src/search.rs
@@ -5,12 +5,10 @@ use std::time::{Duration, Instant};
 use arrayvec::ArrayVec;
 use threadpool::ThreadPool;
 
-use crate::chess::{Color, GameResult, Square};
+use crate::chess::{Color, GameResult, Move, Position, Square};
 use crate::eval;
 use crate::limits::Limits;
 use crate::movepicker::{MAX_MOVES, MovePicker};
-use crate::moves::Move;
-use crate::position::Position;
 use crate::tt::{Entry, EntryType, Table};
 
 pub struct SearchCop {
